Refactor: styles, routes 폴더 생성으로 인한 import 경로 수정
