+++ src/main.tsx
@@ -1,30 +1,30 @@
+import React from "react";
 import ReactDOM from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 import axios from "axios";
+import router from "./routes/Router";
+import "@/styles/global.css";
 import "react-toastify/dist/ReactToastify.css";
-import "./toast.css";
-
-import router from "./Router";
-import "./global.css";
+import "@/styles/toast.css";
 
 axios.defaults.withCredentials = true;
 axios.defaults.baseURL = import.meta.env.VITE_APP_API_ENDPOINT;
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
-  // <React.StrictMode>
-  <>
-    <RouterProvider router={router} />
-    <ToastContainer
-      position="top-center"
-      autoClose={1000}
-      hideProgressBar={false}
-      closeOnClick={true}
-      pauseOnHover={true}
-      draggable={true}
-      theme="light"
-      limit={1}
-    />
-  </>,
-  // </React.StrictMode>,
+  <React.StrictMode>
+    <>
+      <RouterProvider router={router} />
+      <ToastContainer
+        position="top-center"
+        autoClose={1000}
+        hideProgressBar={false}
+        closeOnClick={true}
+        pauseOnHover={true}
+        draggable={true}
+        theme="light"
+        limit={1}
+      />
+    </>
+  </React.StrictMode>,
 );
